Renamed initializers
